% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_data.R
\name{add_feature_data}
\alias{add_feature_data}
\title{Add Featuredata}
\usage{
add_feature_data(object, featuredata, col.name = colnames(featuredata))
}
\arguments{
\item{object}{vistimeseq object}

\item{featuredata}{Data frame where the row names are feature names
and the columns are additional feature attributes}

\item{col.name}{Column names in featuredata to add to the object}
}
\value{
Seurat object where the additional feature data has been added
as columns in object@feature.data
}
\description{
Adds additional data for features to \code{vistimeseq}  object.
Can be any piece of information associated with a feature (e.g. new
featureIDs, prevalence, total counts). The additional data can be used in
plotting functions.
}
\examples{
idx <- sample(1:26, n_features(endoderm_small), replace = TRUE)
random_letters <- data.frame(
  letters = LETTERS[idx],
 row.names = feature_names(endoderm_small)
)
endoderm_small <- add_feature_data(
  object = endoderm_small,
  featuredata = random_letters
  )
head(feature_data(endoderm_small))

}
