% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_lags_to_tc}
\alias{add_lags_to_tc}
\title{Add differences to the time-course data.}
\usage{
add_lags_to_tc(timecourse, lambda)
}
\arguments{
\item{timecourse}{a data matrix or data.frame where each
column corresponds to consecutive time point.}

\item{lambda}{Weights for each lag difference, for time-course data.
Length of \code{lambda} specifies number of lags to include. By default
lag of order one and two are included with coefficients 0.5 and 0.25
respectively.}
}
\value{
a data matrix with added difference lags.
}
\description{
This function add lag difference features
to the time-course data.
}
