% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addLags}
\alias{addLags}
\title{Add differences to the time-course data.}
\usage{
addLags(object, lambda = c(0.5, 0.25))
}
\arguments{
\item{object}{A \code{TimeSeriesExperiment} object}

\item{lambda}{Weights for each lag difference, for time-course data.
Length of \code{lambda} specifies number of lags to include. By default
lag of order one and two are included with coefficients 0.5 and 0.25
respectively.}
}
\value{
Returns \code{TimeSeriesExperiment} object with lags added 
to elements in \code{timeSeries} slot.
}
\description{
This function concatenates lags to time-course
data stored in elements of the \code{timeSeries} slot.
}
\examples{
data("endoderm_small")
endoderm_small <- collapseReplicates(endoderm_small)
endoderm_small <- makeTimeSeries(endoderm_small)
endoderm_small <- addLags(endoderm_small)
head(timeSeries(endoderm_small, "ts"))
head(timeSeries(endoderm_small, "ts_collapsed"))

}
