% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesExperiment-methods.R
\docType{methods}
\name{assays}
\alias{assays}
\alias{assays<-,TimeSeriesExperiment,ANY-method}
\title{Assay data for \code{TimeSeriesExperiment}}
\usage{
\S4method{assays}{TimeSeriesExperiment}(x, ..., withDimnames = TRUE)

\S4method{assays}{TimeSeriesExperiment,ANY}(x, ...,
  withDimnames = TRUE) <- value
}
\arguments{
\item{x}{a \code{TimeSeriesExperiment} object}

\item{...}{argiments to other functions.}

\item{value}{a \link{DataFrame} with new sample information}
}
\value{
a data matrix
}
\description{
Getter and setter methods for the assay slot
}
\examples{
data("endoderm_small")
head(colData(endoderm_small))
newdf <- colData(endoderm_small)
newdf$random <- sample(ncol(endoderm_small), ncol(endoderm_small))
colData(endoderm_small) <- S4Vectors::DataFrame(newdf)
head(colData(endoderm_small))

}
