% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{assign_cluster}
\alias{assign_cluster}
\title{Compute cluster assignment.}
\usage{
assign_cluster(hclst, dynamic = FALSE, h = 0.5, ...)
}
\arguments{
\item{hclst}{an object of class hclust representing the clustering of
features of X.}

\item{dynamic}{whether dynamic branch cutting should be done for cluster
assignment.}

\item{h}{a numeric scalar for the height of the tree if static cutting is
done for cluster assignment.}

\item{...}{other parameters for dynamicTreeCut::cutreeDynamic().}
}
\value{
a matrix of cluster assignments.
}
\description{
This function computes the cluster assignment for hierarchical
clustering results using static or dynamics methods.
}
