% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{assign_cluster_static}
\alias{assign_cluster_static}
\title{Cluster assignment using static brnach cutting.}
\usage{
assign_cluster_static(hclst, h = NULL, k = NULL)
}
\arguments{
\item{hclst}{an object of class \code{hclust} representing the clustering of
features of X.}

\item{h}{a the fraction of the max tree height at which to cut and assign
labels.}

\item{k}{an integer scalar or vector with the desired number of groups}
}
\value{
a data.frame containing cluster assignments.
}
\description{
This function computes the cluster assignment for hierarchical
clustering results using static branch cutting.
}
