% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{assignClusterDynamic}
\alias{assignClusterDynamic}
\title{Cluster assignment using dynamic branch cutting.}
\usage{
assignClusterDynamic(hclst, max_height = 0.9, ...)
}
\arguments{
\item{hclst}{an object of class \code{hclust} representing the clustering of
features of X.}

\item{max_height}{a fraction of the total tree height used to compute
\code{maxTreeHeight} argument for \code{\link[dynamicTreeCut]{cutreeDynamic}}
function from \code{dynamicTreeCut} package.}

\item{...}{other parameters for \code{\link[dynamicTreeCut]{cutreeDynamic}}.}
}
\value{
a data.frame containing cluster assignments.
}
\description{
This function computes the cluster assignment for hierarchical
clustering results using dynamic branch cutting.
}
