% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timevis_class.R
\name{check_timevis}
\alias{check_timevis}
\title{Checks validity of \code{timevis} class.}
\usage{
check_timevis(object)
}
\arguments{
\item{object}{a data object}
}
\value{
\code{TRUE} if valid and a character vector with errors.
}
