% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_class.R
\name{check_vistimeseq}
\alias{check_vistimeseq}
\title{Checks validity of \code{vistimeseq} class.}
\usage{
check_vistimeseq(object)
}
\arguments{
\item{object}{a data object}
}
\value{
\code{TRUE} if valid and a character vector with errors.
}
\description{
function for checking \code{vistimeseq} vaidity
}
