% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cluster_data}
\alias{cluster_data}
\title{Cluster time series data.}
\usage{
cluster_data(X, subset = 1:nrow(X), alpha_coeff = NULL,
  method = "euclidean", ...)
}
\arguments{
\item{X}{a data matrix or data.frame where rows are time series.}

\item{subset}{indices or rownames of rows to keep.}

\item{alpha_coeff}{the coefficients for the added diff/lag features, by default set
as in 'add_diff' function.}

\item{method}{the distance metric for the dissimilarity used for clustering.}

\item{...}{other parameters for 'assign_cluster' function}
}
\value{
a list with the hclust object and the clust.assignment data.frame.
}
\description{
Find the cluster assignment for time series data.
If 'subset' is specified the clsutering is performed on the subset
of the data, and the rest of the rows are assigned based on the distances
to the centroids of computed clusters.
}
