% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_timecourse_features}
\alias{cluster_timecourse_features}
\title{Cluster timecourse features.}
\usage{
cluster_timecourse_features(object, n_top_feat = 1000, groups = "all",
  lambda = c(0.5, 0.25), clust_params = list())
}
\arguments{
\item{object}{A \code{vistimeseq} object}

\item{n_top_feat}{A number of top most variable time-course features to use
for clustering.}

\item{groups}{One or multiple groups from \code{object@group} to take
into account when aggregating time-course features.}

\item{lambda}{Weights for each lag difference, for time-course data.
Length of \code{lambda} specifies number of lags to include. By default
lag of order one and two are included with coefficients 0.5 and 0.25
respectively.}

\item{clust_params}{A list contating arguments for hierarchical clustering.
For details see \code{\link{cluster_data}}.}
}
\value{
a \code{vistimeseq} object with cluster assignment stored
in \code{cluster.map} slot.
}
\description{
Find the cluster assignment for timecourse features.
Clustering computed on top "n_top_feat" features most variable over time
in each of the selected "groups". The cluster assignment of the remaining
genes is based on the distance to the closest cluster centroid previously
obtained. Hierarchical clustering is performed and both static and dynamic
branch cutting algorithm are available for assigning cluster membership.
}
\examples{
endoderm_small
endoderm_small <- cluster_timecourse_features(endoderm_small)
head(get_cluster_map(endoderm_small))

}
