% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesExperiment-methods.R
\docType{methods}
\name{clusterAssignment}
\alias{clusterAssignment}
\alias{clusterAssignment,TimeSeriesExperiment-method}
\alias{clusterMap}
\alias{clusterMap,TimeSeriesExperiment-method}
\title{Cluster analysis results}
\usage{
clusterAssignment(object, ...)

\S4method{clusterAssignment}{TimeSeriesExperiment}(object, name = NULL)

clusterMap(object, ...)

\S4method{clusterMap}{TimeSeriesExperiment}(object)
}
\arguments{
\item{object}{a \code{TimeSeriesExperiment} object.}

\item{...}{argiments to other functions.}

\item{name}{one of elements of 'clusterAssignment' slot: 'settings', 
'hclust', 'cluster_map', 'clust_centroids'. If NULL, all elements are 
returned.}
}
\value{
a \code{data.frame}
}
\description{
Getter methods for \code{clusterAssignment} slot
of a \code{TimeSeriesExperiment} object. The slot is a list with 
with elements named: 'settings', 'hclust', 'cluster_map',
'clust_centroids' storing results from running \link{clusterTimeSeries}
function.
}
\examples{

data("endoderm_small")
endoderm_small <- clusterTimeSeries(endoderm_small)
clusterAssignment(endoderm_small, name = 'settings')
head(clusterAssignment(endoderm_small, name = 'final_cluster_map'))
head(clusterAssignment(endoderm_small, name = 'clust_centroids'))

}
