% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesExperiment-methods.R
\docType{methods}
\name{colData<-,TimeSeriesExperiment,ANY-method}
\alias{colData<-,TimeSeriesExperiment,ANY-method}
\title{Column data for \code{TimeSeriesExperiment}}
\usage{
\S4method{colData}{TimeSeriesExperiment,ANY}(x, ...) <- value
}
\arguments{
\item{x}{a \code{TimeSeriesExperiment} object}

\item{...}{argiments to other functions.}

\item{value}{a \link{DataFrame} with new sample information}
}
\value{
a \link[S4Vectors:DataFrame-class]{S4Vectors::DataFrame}
}
\description{
\code{colData()} slots holds information on individual 
samples including corresponding sample name in column 'sample' as well as 
time, group and replicate.
}
\details{
The setter also updates the information in \code{timepoint}, 
\code{replicate} and \code{group} slots and resets the time-series
analysis results to \code{NULL}.
}
\examples{
data("endoderm_small")
head(colData(endoderm_small))
newdf <- colData(endoderm_small)
newdf$random <- sample(ncol(endoderm_small), ncol(endoderm_small))
colData(endoderm_small) <- newdf
head(colData(endoderm_small))

}
