% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse_replicates}
\alias{collapse_replicates}
\title{Collapse data over replicates.}
\usage{
collapse_replicates(object, FUN = mean)
}
\arguments{
\item{object}{A \code{vistimeseq} object}

\item{FUN}{the aggreagate function. Default is mean}
}
\value{
Returns \code{vistimeseq} object after collapsing
over replicates. Collapsed data is stored \code{sample.data.collapsed}
and \code{data.collapsed} slots.
}
\description{
This function aggregates the data over replicates,
i.e. returns collapse data for each group and at each time point.
}
\examples{
data("endoderm_small")
endoderm_small <- normalize_data(endoderm_small)
endoderm_small <- collapse_replicates(endoderm_small)
head(collapsed_data(endoderm_small))

}
