% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesExperiment-methods.R
\docType{methods}
\name{assayCollapsed}
\alias{assayCollapsed}
\alias{assayCollapsed,TimeSeriesExperiment-method}
\alias{assayCollapsed<-}
\alias{assayCollapsed<-,TimeSeriesExperiment-method}
\alias{colDataCollapsed}
\alias{colDataCollapsed,TimeSeriesExperiment-method}
\alias{colDataCollapsed<-}
\alias{colDataCollapsed<-,TimeSeriesExperiment-method}
\title{Collapsed data.}
\usage{
assayCollapsed(object, ...)

\S4method{assayCollapsed}{TimeSeriesExperiment}(object)

assayCollapsed(object, ...) <- value

\S4method{assayCollapsed}{TimeSeriesExperiment}(object) <- value

colDataCollapsed(object, ...)

\S4method{colDataCollapsed}{TimeSeriesExperiment}(object)

colDataCollapsed(object, ...) <- value

\S4method{colDataCollapsed}{TimeSeriesExperiment}(object) <- value
}
\arguments{
\item{object}{a \code{TimeSeriesExperiment} object.}

\item{...}{argiments to other functions.}

\item{value}{a numerical matrix}
}
\value{
a \link{DataFrame}
}
\description{
Assay and colData collapsed over replicates. The values 
can be computed with set with \link{collapseReplicates} function.
}
\examples{
data("endoderm_small")
endoderm_small <- collapseReplicates(endoderm_small)
head(assayCollapsed(endoderm_small))

data("endoderm_small")
endoderm_small <- collapseReplicates(endoderm_small)
head(colDataCollapsed(endoderm_small))

}
