% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_methods.R
\docType{methods}
\name{collapsed_data}
\alias{collapsed_data}
\alias{collapsed_data,vistimeseq-method}
\title{Accessors for the 'data.collapsed' slot of a \code{vistimeseq} object.}
\usage{
collapsed_data(object)

\S4method{collapsed_data}{vistimeseq}(object)
}
\arguments{
\item{object}{a \code{vistimeseq} object.}
}
\value{
a \code{data.frame}
}
\description{
'data.collapsed' slots holds dataset collapsed by replicates
}
\examples{
data("endoderm_small")
endoderm_small <- collapse_replicates(endoderm_small)
head(collapsed_data(endoderm_small))

}
