% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapseReplicates}
\alias{collapseReplicates}
\title{Collapse data over replicates.}
\usage{
collapseReplicates(object, FUN = mean)
}
\arguments{
\item{object}{A \code{TimeSeriesExperiment} object}

\item{FUN}{the aggreagate function. Default is mean}
}
\value{
Returns \code{TimeSeriesExperiment} object after collapsing
over replicates. Collapsed data is stored \code{sample.data.collapsed}
and \code{data.collapsed} slots.
}
\description{
This function aggregates the data over replicates,
i.e. returns collapse data for each group and at each time point.
}
\examples{
data("endoderm_small")
endoderm_small <- collapseReplicates(endoderm_small)
assayCollapsed(endoderm_small)[1:10, 1:6]

}
