% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_timecourse}
\alias{convert_to_timecourse}
\title{Convert data to time-course.}
\usage{
convert_to_timecourse(object, feature.norm.method = "scale_feat_sum")
}
\arguments{
\item{object}{A \code{timvis} object}

\item{feature.norm.method}{Method for feature normalization.
Currently supports only (default) scaling by total feature
sum, "scale_feat_sum".}
}
\value{
Returns \code{timvis} object after conversion to
time-course format. Converted data is stored in
\code{timecourse.data} slot.
}
\description{
This function converts the wide data matrix
to time-course long \code{data.frame} format where each
row gives data values over time (at each time point) for each
feature, group, and replicate.
}
\examples{
endoderm_small
endoderm_small <- normalize_data(endoderm_small)
endoderm_small <- collapse_replicates(endoderm_small)
endoderm_small <- convert_to_timecourse(endoderm_small)
head(endoderm_small@timecourse.data$tc)

}
