% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_methods.R
\docType{methods}
\name{n_samples}
\alias{n_samples}
\alias{n_samples,vistimeseq-method}
\alias{n_samples}
\alias{n_features}
\alias{n_features,vistimeseq-method}
\alias{n_features}
\title{Number of samples and features}
\usage{
n_samples(object)

\S4method{n_samples}{vistimeseq}(object)

n_features(object)

\S4method{n_features}{vistimeseq}(object)
}
\arguments{
\item{object}{a \code{vistimeseq} object.}
}
\description{
Returns number of samples or in the data.
}
\examples{

endoderm_small
n_samples(endoderm_small)

endoderm_small
n_features(endoderm_small)

}
