% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dataToTimeSeries}
\alias{dataToTimeSeries}
\title{Data to time-series}
\usage{
dataToTimeSeries(X, timepoint, group = NULL, replicate = NULL)
}
\arguments{
\item{X}{a data matrix or data.frame where columns correspond to
samples and rows to features.}

\item{timepoint}{a vector of length equal to ncol(X) indicating the time
variable corresponding to the sample (data column).}

\item{group}{a vector of length equal to ncol(X) indicating the group
membership corresponding to the sample (data column).}

\item{replicate}{a vector of length equal to ncol(X) indicating
the replicate variable corresponding to the sample (data column).}
}
\value{
a data.frame with ordered time series for each replicate.
}
\description{
Function that splits data to time series for each
replicate included.
}
\examples{
X <- matrix(rnorm(1000), ncol = 50)
group <- rep(c("A", "B"), each = 25)
replicate <- rep(paste0("rep", 1:5), each = 5)
time <- rep(1:5, 10)
tc <- dataToTimeSeries(X, time, replicate, group)
head(tc)

}
