% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_methods.R
\docType{methods}
\name{get_diff_expr}
\alias{get_diff_expr}
\alias{get_diff_expr,vistimeseq-method}
\alias{get_diff_expr}
\title{Accessors for the 'diff.expr' slot of a \code{vistimeseq} object.}
\usage{
get_diff_expr(object, type = "all")

\S4method{get_diff_expr}{vistimeseq}(object, type = "all")
}
\arguments{
\item{object}{a \code{vistimeseq} object.}

\item{type}{one of elements of 'dim.red' slot: 'timepoint_de',
'trajectory_de' or 'all' for returning the entire list.}
}
\value{
a \code{data.frame} or a list of \code{data.frame}s
}
\description{
'diff.expr' slots is a list with differential expression results
possibly containing elements named 'timepoint_de', and 'trajectory_de'.
}
\examples{
endoderm_small <- trajectory_de(endoderm_small)
head(get_diff_expr(endoderm_small, "trajectory_de"))

}
