% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesExperiment-methods.R
\docType{methods}
\name{differentialExpression}
\alias{differentialExpression}
\alias{differentialExpression,TimeSeriesExperiment-method}
\title{Differential Expression for \code{TimeSeriesExperiment}}
\usage{
differentialExpression(object, ...)

\S4method{differentialExpression}{TimeSeriesExperiment}(object,
  name = NULL)
}
\arguments{
\item{object}{a \code{TimeSeriesExperiment} object.}

\item{...}{argiments to other functions.}

\item{name}{one of elements of 'differentialExpression' slot: 
'timepoint_de', 'trajectory_de'. If NULL, all elements are returned.}
}
\value{
a \code{data.frame} or a list of \code{data.frame}s
}
\description{
Getter method for \code{differentialExpression} slot 
of a \code{TimeSeriesExperiment}. The slot is a list with differential 
expression results possibly containing elements named 'timepoint_de', and 
'trajectory_de' computed with \link{timepointDE} and \link{trajectoryDE}
functions.
}
\examples{
data("endoderm_small")
endoderm_small <- trajectoryDE(endoderm_small)
head(differentialExpression(endoderm_small, "trajectory_de"))

}
