% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesExperiment-methods.R
\docType{methods}
\name{dimensionReduction}
\alias{dimensionReduction}
\alias{dimensionReduction,TimeSeriesExperiment-method}
\title{Dimension reduction results}
\usage{
dimensionReduction(object, ...)

\S4method{dimensionReduction}{TimeSeriesExperiment}(object, name = NULL)
}
\arguments{
\item{object}{a \code{TimeSeriesExperiment} object.}

\item{...}{argiments to other functions.}

\item{name}{one of elements of 'dimensionReduction' slot: 'pca_sample', 
'pca_feature' and 'pca_eigs' for returning the entire list. If NULL, 
all elements are returned.}
}
\value{
a \code{data.frame} or a list of \code{data.frame}s
}
\description{
Getter methods for \code{dimensionReduction} slot
of a \code{TimeSeriesExperiment} object. The slot is a list of 
\code{data.frames}: 'pca_sample', 'pca_feature' and 'pca_eigs' storing 
results from a PCA projection.
}
\examples{
data("endoderm_small")
endoderm_small <- runPCA(endoderm_small)
head(dimensionReduction(endoderm_small, "pca_sample")[, 1:3])

}
