% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesExperiment-methods.R
\docType{methods}
\name{colnames}
\alias{colnames}
\alias{rownames}
\title{Dimension names getters and  setters for \code{TimeSeriesExperiment} 
object.}
\usage{
\S4method{colnames}{TimeSeriesExperiment}(x) <- value

\S4method{rownames}{TimeSeriesExperiment}(x) <- value
}
\arguments{
\item{x}{a \code{TimeSeriesExperiment} object.}

\item{value}{a character vector or a list of two character
vectors with new dimension names.}
}
\value{
a character vector
}
\description{
\code{colnames()}, \code{rownames()}
can be used to reset the dimensions, column and row names respectively. 
The data will be updated across all slots of \code{TimeSeriesExperiment}.
}
\details{
Setting \code{colnames()} automatically updates
information in  \code{dimensionReduction} slot.

Setting \code{rownames()} automatically updates
information in  \code{timeSeries}, \code{clusterAssignment} 
and \code{dimensionReduction} slots.
}
\examples{
data("endoderm_small")
head(colnames(endoderm_small))
colnames(endoderm_small) <- paste0("Smp", 1:ncol(endoderm_small))
head(colnames(endoderm_small))

data("endoderm_small")
head(rownames(endoderm_small))
rownames(endoderm_small) <- paste0("Feat", 1:nrow(endoderm_small))
head(rownames(endoderm_small))
}
