% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{endoderm}
\alias{endoderm}
\alias{endoderm_small}
\title{Prices of 50,000 round cut diamonds.}
\format{A data frame with 53940 rows and 10 variables}
\source{
\url{http://www.diamondse.info/}
}
\usage{
endoderm

endoderm_small
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds. The variables are as follows:
}
\details{
\describe{
  \item price.
  \item carat.
}
}
\keyword{datasets}
