% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{endoderm_small}
\alias{endoderm_small}
\title{Endoderm differentiation}
\format{
A \code{TimeSeriesExperiment} object
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE98411}
}
\usage{
data(endoderm_small)
}
\description{
A `TimeSeriesExperiment` object contating gene expression data, 
feature and sample data from a study by Blake et al. "A Comparative Study Of 
Endoderm Differentiation In Humans And Chimpanzees". Here we include only 
a small fraction of the data including top 250 most variable genes, to serve
as a toy exapmle for TimeSeriesExperiment.
}
\references{
Blake et al. (2017)
(\href{https://www.biorxiv.org/content/early/2017/05/09/135442}{bioRxiv})
}
\keyword{datasets}
