% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_methods.R
\docType{methods}
\name{get_cluster_map}
\alias{get_cluster_map}
\alias{get_cluster_map,vistimeseq-method}
\alias{feat_clust}
\alias{get_cluster_hclust}
\alias{get_cluster_hclust,vistimeseq-method}
\title{Accessors for the "cluster_map" element of 'cluster.features' slot of
a \code{vistimeseq} object.}
\usage{
get_cluster_map(object)

\S4method{get_cluster_map}{vistimeseq}(object)

get_cluster_hclust(object)

\S4method{get_cluster_hclust}{vistimeseq}(object)
}
\arguments{
\item{object}{a \code{vistimeseq} object.}
}
\value{
a \code{data.frame}

an hclust object
}
\description{
'cluster.features' slots is a list with '"hclust", "cluster_map"
where first is an hclust object and second is a \code{data.frame} with 
cluster assignement, both computed with \code{cluster_timecourse_features()}
function.

'cluster.features' slots is a list with '"hclust", "cluster_map"
where first is an hclust object and second is a \code{data.frame} with
cluster assignement, both computed with \code{cluster_timecourse_features()}
function.
}
\examples{

endoderm_small
endoderm_small <- cluster_timecourse_features(endoderm_small)
head(get_cluster_map(endoderm_small))

endoderm_small
endoderm_small <- cluster_timecourse_features(endoderm_small)
plot(get_cluster_hclust(endoderm_small), labels = FALSE, xlab = "genes",
      sub = "")

}
