% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_methods.R
\docType{methods}
\name{n_features}
\alias{n_features}
\alias{n_features}
\alias{n_features,}
\alias{n_feat}
\title{Number of features}
\usage{
n_features(object)

\S4method{n_features}{vistimeseq}(object)
}
\arguments{
\item{object}{a \code{vistimeseq} object.}
}
\description{
Returns number of features in the data.
}
\examples{

endoderm_small
n_features(endoderm_small)

}
