% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_methods.R
\docType{methods}
\name{feature_names}
\alias{feature_names}
\alias{feature_names<-}
\alias{feature_data}
\alias{feature_data<-}
\alias{feature_names}
\alias{feature_names}
\alias{feature_data}
\alias{feature_data}
\title{Accessors for the 'feature.names' slot of a \code{vistimeseq} object.}
\usage{
feature_names(object)

feature_names(object) <- value

feature_data(object)

feature_data(object) <- value

\S4method{feature_names}{vistimeseq}(object)

\S4method{feature_names}{vistimeseq}(object) <- value

\S4method{feature_data}{vistimeseq}(object)

\S4method{feature_data}{vistimeseq}(object) <- value
}
\arguments{
\item{object}{a \code{vistimeseq} object.}

\item{object}{a \code{vistimeseq} object.}
}
\description{
'feature.names' slots holds feature names

'feature.data' slots holds information on individual features including
corresponding feature name in column 'feature'.
}
\examples{

endoderm_small
head(feature_names(endoderm_small))


endoderm_small
head(feature_data(endoderm_small))

}
