% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_methods.R
\docType{methods}
\name{feature_names}
\alias{feature_names}
\alias{feature_names,vistimeseq-method}
\alias{feature_names}
\alias{feature_names<-}
\alias{feature_names<-,vistimeseq-method}
\title{Accessors for the 'feature.names' slot of a \code{vistimeseq} object.}
\usage{
feature_names(object)

\S4method{feature_names}{vistimeseq}(object)

feature_names(object) <- value

\S4method{feature_names}{vistimeseq}(object) <- value
}
\arguments{
\item{object}{a \code{vistimeseq} object.}

\item{value}{a character vectore with new feature names}
}
\value{
a character vector
}
\description{
'feature.names' slots holds feature names
}
\examples{
endoderm_small
head(feature_names(endoderm_small))

}
