% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_data.R
\name{filter_features}
\alias{filter_features}
\title{Filter features}
\usage{
filter_features(object, features)
}
\arguments{
\item{object}{vistimeseq object}

\item{features}{features (genes) to keep}
}
\value{
"vistimeseq" object
}
\description{
Filters "vistimeseq" object to keep only chosen features.
All relevant slots are updates.
}
\examples{
abc <- sample(1:10)

}
