% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterFeatures}
\alias{filterFeatures}
\title{Filter features}
\usage{
filterFeatures(object, features)
}
\arguments{
\item{object}{TimeSeriesExperiment object}

\item{features}{features (genes) to keep}
}
\value{
"TimeSeriesExperiment" object
}
\description{
Filters \code{TimeSeriesExperiment} object to keep only chosen 
features. All relevant slots are updates.
}
\details{
The slots for collapsed data and time series formatted data are
filtered accordingly, but \code{dimensionReduction}, 
\code{clusterAssignment} and \code{differentialExpression} are reset to 
\code{NULL} as different set of features would output in different results.
}
\examples{
data("endoderm_small")
features <- 1:100
endoderm_small <- filterFeatures(endoderm_small, features)
endoderm_small

}
