% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sig_genes}
\alias{get_sig_genes}
\title{Venn diagram for features significant at each timpoints}
\usage{
get_sig_genes(df, timepoint, feat_data = NULL)
}
\arguments{
\item{df}{a data.frame of features by timepoints listing significant
features for each timepoint, NA entries stands for non-significant,
features do not have to be in the same order.}

\item{timepoint}{a timepoint of interest (must be one of colnames(df))}

\item{feat_data}{a data.frame contating gene data.}
}
\value{
a vector of significant features or a data.frame with accompanying data.
}
