% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_methods.R
\docType{methods}
\name{get_group}
\alias{get_group}
\alias{set_group<-}
\alias{get_group}
\alias{set_group}
\title{Accessors for the 'group' slot of a \code{vistimeseq} object.}
\usage{
get_group(object)

set_group(object) <- value

\S4method{get_group}{vistimeseq}(object)

\S4method{set_group}{vistimeseq}(object) <- value
}
\arguments{
\item{object}{a \code{vistimeseq} object.}
}
\description{
'grop' slots holds feature names
}
\examples{

endoderm_small
head(get_group(endoderm_small))

}
