% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesExperiment-methods.R
\docType{methods}
\name{groups}
\alias{groups}
\alias{groups,TimeSeriesExperiment-method}
\alias{groups<-}
\alias{groups<-,TimeSeriesExperiment-method}
\title{Group information}
\usage{
groups(object, ...)

\S4method{groups}{TimeSeriesExperiment}(object)

groups(object, ...) <- value

\S4method{groups}{TimeSeriesExperiment}(object) <- value
}
\arguments{
\item{object}{a \code{TimeSeriesExperiment} object.}

\item{...}{argiments to other functions.}

\item{value}{a character vector with new group membership}
}
\value{
a character vector
}
\description{
Group information
}
\examples{
data("endoderm_small")
head(groups(endoderm_small))
groups(endoderm_small) <- sample(c("A", "B"), 
    ncol(endoderm_small), replace = TRUE)
head(groups(endoderm_small))

}
