% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_class.R
\name{impute_data}
\alias{impute_data}
\title{Imputing missing samples}
\usage{
impute_data(raw.data, sample.data)
}
\arguments{
\item{raw.data}{Raw input data with columns corresponding to samples
(observations) and rows to features}

\item{sample.data}{Optional. A \code{data.frame} object were rows are
samples (observations) and columns are sample attributes
(e.g. group/condition, replicate, time)}

\item{project}{Project name (string)}
}
\value{
list of modified raw.data and sample.data
}
\description{
Sometimes datasets include a missing sample. Here we impute
the expression values by taking the mean expression of two samples
from the same group and replicate (same individual) at two times surrounding
the time of the missing sample
}
