% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join_sources}
\alias{join_sources}
\title{Join data soures.}
\usage{
join_sources(X, sample_data, cluster)
}
\arguments{
\item{X}{a data matrix or data.frame where columns correspond to samples and rows to features.}

\item{sample_data}{a sample data matrix.}

\item{cluster}{a data.frame with features and their cluster assignments.}
}
\value{
a data matrix with added difference lags.
}
\description{
This function add lag difference features to the time series data.
}
