% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeTimeSeries}
\alias{makeTimeSeries}
\title{Convert data to time-series}
\usage{
makeTimeSeries(object, feature.trans.method = "var_stab",
  var.stabilize.method = "asinh")
}
\arguments{
\item{object}{A \code{TimeSeriesExperiment} object}

\item{feature.trans.method}{Method for feature normalization. 
Default "none". Currently supports only "none" (no transformation),
"scale_feat_sum" (scaling by feature sum),
or "var_stab" (variance stabilization). Default is "var_stab".}

\item{var.stabilize.method}{Method for variance stabilization (VST).
Currently, supports "none" (no VST), "log1p" (log plus one), "asinh"
(inverse hyperbolic sine) or "deseq"
(\code{\link[DESeq2]{varianceStabilizingTransformation}} function from
\code{DESeq2} package). Default is "log1p".}
}
\value{
Returns \code{TimeSeriesExperiment} object after conversion to
time-course format. Converted data is stored in
\code{timecourse.data} slot.
}
\description{
This function converts the wide data matrix
to time-course long \code{data.frame} format where each
row gives data values over time (at each time point) for each
feature, group, and replicate.
}
\examples{
data("endoderm_small") 
endoderm_small <- makeTimeSeries(endoderm_small)
names(timeSeries(endoderm_small))
head(timeSeries(endoderm_small)[[1]])

}
