% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesExperiment-class.R
\name{makeTimeSeriesExperimentFromExpressionSet}
\alias{makeTimeSeriesExperimentFromExpressionSet}
\title{TimeSeriesExperiment constructor from ExpressionSet}
\usage{
makeTimeSeriesExperimentFromExpressionSet(eset, timepoint = NULL,
  group = NULL, replicate = NULL)
}
\arguments{
\item{eset}{ExpressionSet object}

\item{timepoint}{a vector indicating timepoint at which each sample 
was collected or a character string equal to one of the column names 
of a supplied \code{ExpressionSet}.}

\item{group}{a vector indicating a group membership for each sample 
or a character string equal to one of the column names of a supplied 
\code{ExpressionSet} If not specified, the group is set to 'G1' for
each sample.}

\item{replicate}{a vector indicating a replicate id of each sample 
or a character string equal to one of the column names of a supplied 
\code{ExpressionSet} If not specified, the replicate is set to 'R1' 
for each sample.}
}
\value{
Returns an initialized TimeSeriesExperiment object.
}
\description{
\code{TimeSeriesExperiment} constructor initializes the 
\code{TimeSeriesExperiment} object from 
\code{ExpressionSet}
and populates the time, replicate, and group slots.
}
\details{
\code{TimeSeriesExperiment} is an extension of 
\code{SummarizedExperiment} class.
}
\examples{
raw <- matrix(runif(3000), ncol = 30)
pheno.data <- data.frame(
   time = rep(rep(1:5, each = 3), 2),
   replicate = rep(1:3, 10),
   group = rep(1:2, each = 15))
feature.data <- data.frame(
   feature = paste0("F", 1:100)
)
\dontrun{
    library(Biobase)
    test_eset <- ExpressionSet(
         raw, phenoData = AnnotatedDataFrame(pheno.data),
         featureData = AnnotatedDataFrame(feature.data))
    test_TimeSeriesExperiment <- makeTimeSeriesExperimentFromExpressionSet(
       test_eset, timepoint = "time", group = "group",
       replicate = "replicate")
    test_TimeSeriesExperiment
}


}
