% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{meltMatrix}
\alias{meltMatrix}
\title{Melt data matrix.}
\usage{
meltMatrix(X)
}
\arguments{
\item{X}{a data matrix.}
}
\value{
a data matrix in a long format.
}
\description{
Convert data matrix to a long format
}
\examples{
Z <- matrix(rnorm(100), 20)
Z.m <- meltMatrix(Z)
head(Z.m)

}
