% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize Data}
\usage{
normalize_data(object, sample.norm.method = "scale_common_factor",
  column.scale.factor = 1e+06)
}
\arguments{
\item{object}{A \code{timvis} object or a data matrix/data frame.}

\item{sample.norm.method}{Method for sample normalization.
Currently supports only scaling to a common factor,
"scale_common_factor" which with  \code{column.scale.factor} = 1e+06
is equivalent to CPM normalization.}

\item{column.scale.factor}{Sets the scale factor for sample-level
normalization}
}
\value{
Returns \code{timvis} object after normalization.
Normalized data is stored \code{data} slot.
}
\description{
Normalize data the assay data.
}
\examples{
endoderm_small
endoderm_small <- normalize_data(endoderm_small)
head(endoderm_small@data)

}
