% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_over_representation.R
\name{pathway_enrichment}
\alias{pathway_enrichment}
\title{Pathway enrichment testing.}
\usage{
pathway_enrichment(object, features, species, clustered = TRUE,
  kegg = FALSE, fltr_DE = 0.1, fltr_N = 500, fltr_P.DE = 0.05, ...)
}
\arguments{
\item{object}{A \code{vistimeseq} object.}

\item{features}{A vector of names of selected features to plot.}

\item{species}{A character string  specifying the species.
See \code{\link[limma]{goana}} for details.}

\item{clustered}{Whether \code{features} should be grouped based on
cluster assignment (stored in \code{object@cluster.features$cluster_map})
and tested as separate sets.}

\item{kegg}{Whether KEGG pathways should be used instead of gene ontology
(GO)}

\item{fltr_DE}{A scalar fraction of the number of genes in tested set
to use as a threshold for filtering genes based on "DE" column
(the number of genes in the DE set). Default is 0.1, i.e. at least
0.1 genes in the set must be present in the pathway.}

\item{fltr_N}{A number of genes used as a threshold to filter out all
pathway terms of size greater than the threshold. Default is 500.}

\item{fltr_P.DE}{A p-value threshold to filter out terms in the enrichment
results. Default is 0.05.}

\item{...}{other parameters for \code{\link[limma]{goana}} or
\code{\link[limma]{kegga}}.}
}
\value{
a \code{data.frame} or list of \code{data.frame}s with enrichment
results.
}
\description{
This is a wrapper around \code{\link[limma]{goana}} function
from \code{limma} package for testing over-representation of gene ontology
(GO) terms or KEGG pathways in sets of genes.
}
\examples{
\dontrun{
endoderm
endoderm <- normalize_data(endoderm)
endoderm <- trajectory_de(endoderm)
genes_with_de_trajectory <- get_diff_expr(endoderm, "trajectory_de") \%>\%
  filter(pval <= max(0.05, min(pval)), R2 > 0.7) \%>\%
  arrange(-R2)
res <- pathway_enrichment(
  object = endoderm, clustered = FALSE,
  features = genes_with_de_trajectory$feature,
  species = "Hs", fltr_DE = 0, fltr_N = Inf, fltr_P.DE = 0.05)
head(res)
}

}
