% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_over_representation.R
\name{pathwayEnrichment}
\alias{pathwayEnrichment}
\title{Pathway enrichment testing.}
\usage{
pathwayEnrichment(
  object,
  features,
  species,
  feature_column = "feature",
  universe = NULL,
  clustered = TRUE,
  kegg = FALSE,
  ontology = c("BP", "CC", "MF"),
  fltr_DE = 0.1,
  fltr_N = 500,
  fltr_P.DE = 0.05,
  ...
)
}
\arguments{
\item{object}{A \code{TimeSeriesExperiment} object.}

\item{features}{A vector of ENTREZID for enrichment testing.}

\item{species}{A character string    specifying the species.
See \code{\link[limma]{goana}} for details.}

\item{feature_column}{the feature column in 'feature.data' slot holding
ENTREZIDs, by deafult the 'feature' column.}

\item{universe}{A vector of genes in the universe. By default
all the genes in the 'raw.data' slot.}

\item{clustered}{Whether \code{features} should be grouped based on
cluster assignment (stored in \code{object@cluster.features$cluster_map})
and tested as separate sets.}

\item{kegg}{Whether KEGG pathways should be used instead of gene ontology
(GO)}

\item{ontology}{character vector of ontologies to be included in output.
Elements should be one or more of "BP", "CC" or "MF".}

\item{fltr_DE}{A scalar fraction of the number of genes in tested set
to use as a threshold for filtering genes based on "DE" column
(the number of genes in the DE set). Default is 0.1, i.e. at least
0.1 genes in the set must be present in the pathway.}

\item{fltr_N}{A number of genes used as a threshold to filter out all
pathway terms of size greater than the threshold. Default is 500.}

\item{fltr_P.DE}{A p-value threshold to filter out terms in the enrichment
results. Default is 0.05.}

\item{...}{other parameters for \code{\link[limma]{goana}} or
\code{\link[limma:goana]{limma::kegga()}}.}
}
\value{
a \code{data.frame} or list of \code{data.frame}s with enrichment
results.
}
\description{
This is a wrapper around \code{\link[limma]{goana}} function
from \code{limma} package for testing over-representation of gene ontology
(GO) terms or KEGG pathways in sets of genes.
}
\examples{
data("endoderm_small")
selected_genes <- c('114299', '2825', '3855', '221400', '7941',
                    '6164', '1292', '6161', '6144', '23521')
enrich_res <- pathwayEnrichment(
  object = endoderm_small, clustered = FALSE,
  features = selected_genes,
  species = "Hs", ontology = "BP", fltr_DE = 0,
  fltr_N = Inf, fltr_P.DE = 0.05)
head(enrich_res)

}
