% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Plot enrichment results.}
\usage{
plot_enrichment(enrich, n_max = 15)
}
\arguments{
\item{enrich}{a data matrix or data.frame with enrichment result -
outputs from \code{\link{pathway_enrichment}} function or
\code{\link[limma]{goana}}, \code{\link[limma]{kegga}}.
Must contain columns Term, DE, and P.DE.}

\item{n_max}{max number of terms to show}
}
\value{
a ggplot object
}
\description{
Plotting top most enriched terms found with DE methods
and tested for overrepresentation in GO/KEGG db using
goana/kegga from limma package.
}
\examples{
\dontrun{
endoderm_small <- normalize_data(endoderm_small)
endoderm_small <- trajectory_de(endoderm_small)
genes_with_de_trajectory <-
  get_diff_expr(endoderm_small, "trajectory_de") \%>\%
    dplyr::filter(pval <= max(0.05, min(pval)), R2 > 0.7) \%>\%
    dplyr::arrange(-R2)
enrich_res <- pathway_enrichment(
  object = endoderm_small, clustered = FALSE,
  features = feature_names(endoderm_small)[1:10],
  species = "Hs", ontology ="BP")
  plot_enrichment( enrich = enrich_res, n_max = 15)
}

}
