% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_goana}
\alias{plot_goana}
\title{Plot Gene Ontology limma::goana enrichment results.}
\usage{
plot_goana(DF, nGSmax = 15)
}
\arguments{
\item{DF}{a data matrix or data.frame results from limma::goana
with gene list. Must contain columns Term, DE, and P.DE.}
}
\value{
a ggplot object
}
