% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a heatmap}
\usage{
plot_heatmap(object, num.feat = 200, scale = TRUE, feat_desc = "feature",
  sample_desc = "sample")
}
\arguments{
\item{object}{A \code{vistimeseq} object}

\item{num.feat}{Number of top most features to use.}

\item{scale}{Whether to scale the data (by features) before plotting.}

\item{feat_desc}{One of the column names from \code{object@feature.data}
to describe the features.}

\item{sample_desc}{ne of the column names from \code{object@sample.data}
to describe the samples.}

\item{...}{other parameters ggplot.}
}
\value{
Returns a \code{ggplot2} objet.
}
\description{
Generates a heatmap of expression data.
}
\examples{
endoderm_small
endoderm_small <- normalize_data(endoderm_small)
plot_heatmap(endoderm_small)

}
