% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_sample_pca}
\alias{plot_sample_pca}
\title{Plot a standard PCA}
\usage{
plot_sample_pca(object, axis = c(1, 2), col.var = NULL, ...)
}
\arguments{
\item{object}{A \code{vistimeseq} object}

\item{axis}{An integer vector indicating principal components to use for
plotting, by default 1:2.}

\item{col.var}{A character string indicating a column
from sample_data(object) which should be used for coloring
the points. By default NULL.}

\item{...}{other parameters ggplot.}
}
\value{
Returns a \code{ggplot2} objet.
}
\description{
Generates a standard PCA plot of observations in the dataset.
}
\examples{
data("endoderm_small")
endoderm_small <- run_pca(endoderm_small)
plot_sample_pca(endoderm_small, col.var = "group")

}
