% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Plot selected time series.}
\usage{
plot_time_series(X, time, keep_rows = rownames(X), group = NULL,
  replicate = NULL, feature_data = NULL, ncol = 5, smooth = TRUE)
}
\arguments{
\item{X}{a data matrix or data.frame}

\item{time}{a vector of length equal to ncol(X) indicating the time
variable corresponding to the sample (data column).}

\item{keep_rows}{slected features to plot.}

\item{group}{a vector of length equal to ncol(X) indicating
the group membership of the sample.}

\item{replicate}{a vector of length equal to ncol(X) indicating
the replicate (e.g. individual).}

\item{feature_data}{a data.frame contating the information on features.}

\item{ncol}{an integer indicating the number of columns for facetting}
}
\value{
list of ggplot objects
}
