% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_ts_clusters}
\alias{plot_ts_clusters}
\title{Plot (time) series over clusters.}
\usage{
plot_ts_clusters(X, cluster, time, group = NULL, freq_df = NULL,
  plot = TRUE, facet = TRUE, textsize = 5)
}
\arguments{
\item{X}{a data matrix or data.frame}

\item{cluster}{a data.frame containing 'feature' and 'cluster' columns.}

\item{time}{a vector of length equal to ncol(X) indicating the time
variable corresponding to the sample (data column).}

\item{group}{a vector of length equal to ncol(X) indicating
the group membership of the sample.}

\item{freq.df}{a data.frame contating the 'cluster' and 'freq' columns
which indicates the count of genes in each cluster.}
}
\value{
list of ggplot objects
}
