% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_ts_pca}
\alias{plot_ts_pca}
\title{Overlay (time) series over PCA grid}
\usage{
plot_ts_pca(object, axis = 1:2, m = 20, n = 20, group.highlight = NULL,
  linecol = NULL, ...)
}
\arguments{
\item{object}{A \code{vistimeseq} object.}

\item{axis}{An integer vector indicating principal components to use for
plotting, by default 1:2.}

\item{m}{a number of tiles in a grid in the horizontal direction.}

\item{n}{a number of tiles in a grid in the vertical direction.}

\item{group.highlight}{An optional character string indicating the group
subset for which the time-course trends should be plotted. By default all
time-course trends are plotted for all groups.}

\item{linecol}{a vector indicating the color of the gene profile trend line,
different for each group.}

\item{...}{other parameters for the line plots.}
}
\examples{
endoderm_small
endoderm_small <- run_pca(endoderm_small)
plot_ts_pca(endoderm_small)

}
