% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot a heatmap}
\usage{
plotHeatmap(object, num.feat = 200, scale = TRUE,
  feat_desc = "feature", sample_desc = "sample")
}
\arguments{
\item{object}{A \code{TimeSeriesExperiment} object}

\item{num.feat}{Number of top most features to use.}

\item{scale}{Whether to scale the data (by features) before plotting.}

\item{feat_desc}{One of the column names from \code{feature_data(object)}
to describe the features.}

\item{sample_desc}{ne of the column names from \code{sample_data(object)}
to describe the samples.}

\item{...}{other parameters ggplot.}
}
\value{
Returns a \code{ggplot2} objet.
}
\description{
Generates a heatmap of expression data.
}
\examples{
data("endoderm_small")
endoderm_small <- normalizeData(endoderm_small)
\dontrun{
    plotHeatmap(endoderm_small)
}
}
