% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Plot selected time series.}
\usage{
plotTimeSeries(object, features = rownames(object), trans = FALSE,
  smooth = TRUE, ncol = 5, scales = "free")
}
\arguments{
\item{object}{A \code{TimeSeriesExperiment} object.}

\item{features}{A vector of names of selected features to plot.}

\item{trans}{A boolean indicating whether (TRUE) transformed, variance 
stabilized, assay values should be printed or (FALSE) just normalized by
sample depth.}

\item{smooth}{If TRUE a smoothed line is plotted for each gene
and each group, else a piecewise linear average (over replicates) curve
is plotted.}

\item{ncol}{An integer indicating the number of columns for facetting.
Default is 5.}

\item{scales}{character scalar indecating facet scales, by default "free".}
}
\value{
list of ggplot objects
}
\description{
Plotting expression over time for selected genes curve
and colors correspond to distinct groups.
}
\examples{
data("endoderm_small")
feat_to_plot <- rownames(endoderm_small)[1:10]
plotTimeSeries(endoderm_small, features = feat_to_plot, smooth = FALSE)

}
