% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTimeSeriesClusters}
\alias{plotTimeSeriesClusters}
\title{Plot (time) series over clusters.}
\usage{
plotTimeSeriesClusters(object, features = NULL, transparency = 0.5,
  ncol = 4, scales = "free")
}
\arguments{
\item{object}{A \code{TimeSeriesExperiment} object.}

\item{features}{A vector of names of selected features to plot.}

\item{transparency}{transparency of trajectory lines.}

\item{ncol}{number of columns in the factet plot.}

\item{scales}{character scalar indecating facet scales, by default "free".}
}
\value{
ggplot object
}
\description{
Plots timecourse (aggregated over replicates) feature data
faceted by computed clusters and experimental group.
}
\examples{
data("endoderm_small")
endoderm_small <- clusterTimeSeries(endoderm_small)
plotTimeSeriesClusters(endoderm_small)

}
