% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_methods.R
\docType{methods}
\name{project_name}
\alias{project_name}
\alias{project_name}
\title{Project name}
\usage{
project_name(object)

\S4method{project_name}{vistimeseq}(object)
}
\arguments{
\item{object}{a \code{vistimeseq} object.}
}
\description{
Returns project name.
}
\examples{

endoderm_small
project_name(endoderm_small)

}
