% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_methods.R
\docType{methods}
\name{set_replicate<-}
\alias{set_replicate<-}
\alias{get_replicate}
\alias{get_replicate,vistimeseq-method}
\alias{set_replicate<-,vistimeseq-method}
\title{Accessors for the 'replicate' slot of a \code{vistimeseq} object.}
\usage{
set_replicate(object) <- value

get_replicate(object)

\S4method{get_replicate}{vistimeseq}(object)

\S4method{set_replicate}{vistimeseq}(object) <- value
}
\arguments{
\item{object}{a \code{vistimeseq} object.}

\item{value}{a character vector with new replicate ids.}
}
\value{
a character vector
}
\description{
'grop' slots holds feature names
}
\examples{
data("endoderm_small")
head(get_replicate(endoderm_small))

}
