% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesExperiment-methods.R
\docType{methods}
\name{replicates}
\alias{replicates}
\alias{replicates,TimeSeriesExperiment-method}
\alias{replicates<-}
\alias{replicates<-,TimeSeriesExperiment-method}
\title{Replicate information}
\usage{
replicates(object, ...)

\S4method{replicates}{TimeSeriesExperiment}(object)

replicates(object, ...) <- value

\S4method{replicates}{TimeSeriesExperiment}(object) <- value
}
\arguments{
\item{object}{a \code{TimeSeriesExperiment} object.}

\item{...}{argiments to other functions.}

\item{value}{a character vector with new replicate ids.}
}
\value{
a character vector
}
\description{
Replicate information
}
\examples{
data("endoderm_small")
head(replicates(endoderm_small))
replicates(endoderm_small) <- sample(c("R1", "R2", "R3"), 
    ncol(endoderm_small), replace = TRUE)
head(replicates(endoderm_small))
}
