% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesExperiment-methods.R
\docType{methods}
\name{rowData<-,TimeSeriesExperiment-method}
\alias{rowData<-,TimeSeriesExperiment-method}
\title{Row data for \code{TimeSeriesExperiment}}
\usage{
\S4method{rowData}{TimeSeriesExperiment}(x, ...) <- value
}
\arguments{
\item{x}{a \code{TimeSeriesExperiment} object.}

\item{...}{argiments to other functions.}

\item{value}{a \code{DataFrame} with new feature data}
}
\value{
a \link[S4Vectors:DataFrame-class]{S4Vectors::DataFrame}
}
\description{
\code{rowData()} holds information on individual features 
including corresponding feature name in column 'feature'.
}
\examples{
data("endoderm_small")
head(rowData(endoderm_small))
rowData(endoderm_small) <- data.frame(
    feature = rowData(endoderm_small)$feature,
    random = sample(nrow(endoderm_small), nrow(endoderm_small)))
head(rowData(endoderm_small))

}
