% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{run_pca}
\alias{run_pca}
\title{Run PCA scores.}
\usage{
run_pca(object, type, group.selected = NULL, var.stabilize.method = "log",
  pseudocount = 1, log.base = exp(1))
}
\arguments{
\item{object}{A \code{timevis} object}

\item{type}{Either "sample" or "feature" indicating whether to compute PCA
for observations or for time series features.}

\item{group.selected}{An optional character string indicating which group
to subset the data to. By default all groups are considered. Only considered
for \code{type} = "feature" case.}

\item{var.stabilize.method}{Method for variance stabilization (VST)
if computing PCA for \code{type} = "sample". Currently, supports log plus
one normalization, "log", or DESeq2::varianceStabilizingTransformation,
"deseq". By default "lognorm" VST is performed.}

\item{pseudocount}{A numerical indicating the pseudocount in log
transformation}

\item{log.base}{A numerical indicating the base for log
transformation}
}
\value{
Returns \code{timevis} object with PCA results in the \code{dim.red}
slot. The PCA results are lists of 'pca.scores' and 'pca.eigs', and will
enter elements of \code{dim.red} slot named "pca_samples" or "pca_features".
}
\description{
Compute the projection of either samples or time series
features onto a PCA space. In case, of PCA for features, the PCA can be
computed for individual sample group as indicated or for samples from all
groups. In either, case the data is first collapsed over replicates, so
that each gene is represented as a vector of a single time course.
}
\examples{
endoderm_small
endoderm_small <- normalize_data(endoderm_small)
endoderm_small <- run_pca(endoderm_small, type = "sample")
head(endoderm_small@dim.red$pca_sample[["pca.scores"]][, 1:5])
head(endoderm_small@dim.red$pca_sample[["pca.eigs"]])

}
