% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_methods.R
\docType{methods}
\name{collapsed_sample_data}
\alias{collapsed_sample_data}
\alias{collapsed_sample_data,vistimeseq-method}
\alias{collapsed_sample_data}
\title{Accessors for the 'data.collapsed' slot of a \code{vistimeseq} object.}
\usage{
collapsed_sample_data(object)

\S4method{collapsed_sample_data}{vistimeseq}(object)
}
\arguments{
\item{object}{a \code{vistimeseq} object.}
}
\description{
'sample.data.collapsed' slots holds sample data for data
collapsed by replicates
}
\examples{
endoderm_small
endoderm_small <- collapse_replicates(endoderm_small)
head(collapsed_sample_data(endoderm_small))

}
