% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_methods.R
\docType{methods}
\name{n_samples}
\alias{n_samples}
\alias{sample_names}
\alias{sample_names<-}
\alias{sample_data}
\alias{sample_data<-}
\alias{n_samples}
\alias{sample_names}
\alias{sample_names}
\alias{sample_data}
\alias{sample_data}
\title{Number of samples}
\usage{
n_samples(object)

sample_names(object)

sample_names(object) <- value

sample_data(object)

sample_data(object) <- value

\S4method{n_samples}{vistimeseq}(object)

\S4method{sample_names}{vistimeseq}(object)

\S4method{sample_names}{vistimeseq}(object) <- value

\S4method{sample_data}{vistimeseq}(object)

\S4method{sample_data}{vistimeseq}(object) <- value
}
\arguments{
\item{object}{a \code{vistimeseq} object.}

\item{object}{a \code{vistimeseq} object.}

\item{object}{a \code{vistimeseq} object}

\item{collapsed}{whether return sample information for data with replicates
aggregated.}
}
\description{
Returns number of samples in the data.

'sample.names' slots holds sample names

'sample.data' slots holds information on individual samples including
corresponding sample name in column 'sample' as well as time, group
and replicate.
}
\examples{

endoderm_small
n_samples(endoderm_small)


endoderm_small
head(sample_names(endoderm_small))


endoderm_small
head(sample_data(endoderm_small))

}
