% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_methods.R
\docType{methods}
\name{sample_names}
\alias{sample_names}
\alias{sample_names,vistimeseq-method}
\alias{sample_names}
\alias{sample_names<-}
\alias{sample_names<-,vistimeseq-method}
\title{Accessors for the 'sample.names' slot of a \code{vistimeseq} object.}
\usage{
sample_names(object)

\S4method{sample_names}{vistimeseq}(object)

sample_names(object) <- value

\S4method{sample_names}{vistimeseq}(object) <- value
}
\arguments{
\item{object}{a \code{vistimeseq} object.}

\item{value}{a character vector with new sample names}
}
\description{
'sample.names' slots holds sample names
}
\examples{
endoderm_small
head(sample_names(endoderm_small))

}
