% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_methods.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,vistimeseq-method}
\title{show method for \code{vistimeseq}}
\usage{
\S4method{show}{vistimeseq}(object)
}
\arguments{
\item{object}{A vistimeseq object}
}
\value{
nothing, just prints to console
}
\description{
show method for \code{vistimeseq}
}
