% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{static_diff_expr}
\alias{static_diff_expr}
\title{Static differential expression testing.}
\usage{
static_diff_expr(X, sample_data, gene_data = NULL, alpha = 0.05,
  group_var = NULL, test_coeff = NULL)
}
\arguments{
\item{X}{a data matrix or data.frame where rows are time series.}

\item{sample_data}{a data.frame of sample data.}

\item{alpha}{a scalar for level of significance.}

\item{group_var}{a character scalar (one of the column names in
'sample_data') indicating the group sample variable for which DE is tested.}

\item{test_coeff}{a character scalar for test coefficient.}
}
\value{
a list with the hclust object and the clust.assignment data.frame.
}
\description{
This is a wrapper around 'limma' + 'voom' functions for testing
differential expression. We are performing tests for each single timepoint.
}
