% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_methods.R
\docType{methods}
\name{time_course}
\alias{time_course}
\alias{time_course,vistimeseq-method}
\alias{time_course}
\title{Accessors for the 'timecourse.data' slot of a \code{vistimeseq} object.}
\usage{
time_course(object, collapsed = FALSE)

\S4method{time_course}{vistimeseq}(object, collapsed = FALSE)
}
\arguments{
\item{object}{a \code{vistimeseq} object.}

\item{collapsed}{whether collapsed time-course should be returned}
}
\value{
a \code{data.frame}
}
\description{
'timecourse.data' slots is a list with 'tc' and (optionally) 'tc_collapsed'
storing timce-course format of the data.
}
\examples{
endoderm_small
endoderm_small <- convert_to_timecourse(endoderm_small)
head(time_course(endoderm_small))
head(time_course(endoderm_small, collapsed = TRUE))

}
