% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{time_series_diff}
\alias{time_series_diff}
\title{Differences in expression profiles.}
\usage{
time_series_diff(X, time, replicate, group = NULL, reference_name = NULL,
  test_group_name = NULL, alpha_coeff = c(0.5, 0.25))
}
\arguments{
\item{X}{a data matrix or data.frame where rows are time series.}

\item{time}{a vector of length equal to ncol(X) indicating the time
variable corresponding to the sample (data column).}

\item{replicate}{a vector of length equal to ncol(X) indicating
the sample replicate (e.g. individual).}

\item{group}{a vector of length equal to ncol(X) indicating
the group membership of the sample.}

\item{reference_name}{a character scalar indicating the reference group.}

\item{test_group_name}{a character scalar indicating the test group.}

\item{alpha_coeff}{the weights for each lag difference, by default
0.5^(1:(ncol(X)-1))}
}
\value{
a data.frame with columns 'gene', 'diff.mean', 'diff.sd',
'ref.diff.mean', 'ref.diff.sd', 'test.diff.mean', 'test.diff.sd', indicating
respectively the mean and sd of the differences between time series
gene expression profiles of replicates between two groups, within the
reference, and within the test group.
}
\description{
This is function computes differences in gene
expression profiles between two selected groups.
}
