% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{timepoint_de}
\alias{timepoint_de}
\title{Differential timepoint expression testing.}
\usage{
timepoint_de(object, timepoints = "all", min_gene_sum = 1,
  alpha = 0.05)
}
\arguments{
\item{object}{A \code{vistimeseq} object.}

\item{timepoints}{Vector of timepoints to test at.}

\item{min_gene_sum}{A scalar for filtering sparse genes before DE testing.
Default is 1.}

\item{alpha}{A scalar for level of significance. Default is 0.05.}
}
\value{
a \code{vistimeseq} object with timepoint differential expression
testing results stored in 'timepoint_de' element in \code{diff.expr} slot.
}
\description{
This is a wrapper around \code{\link[limma]{limma}}
and \code{\link[limma]{voom}} functions from \code{limma} package
for testing differential expression at specified timepoints.
}
\examples{
endoderm_small <- timepoint_de(endoderm_small, timepoint = 1.0)
head(get_diff_expr(endoderm_small, "timepoint_de")$`1`)

}
