% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{timepoint_venn}
\alias{timepoint_venn}
\title{Venn diagram for features significant at each timpoints}
\usage{
timepoint_venn(df, size = 50, transparency = 0.5, cexil = 0.9,
  cexsn = 1, zcolor = viridis(length(venn.lst)), ...)
}
\arguments{
\item{df}{a data.frame of features by timepoints listing significant
features for each timepoint, NA entries stands for non-significant,
features do not have to be in the same order.}

\item{size}{plot size, in centimeters.}

\item{transparency}{transparency for the color(s) specified with zcolor}

\item{cexil}{a character expansion for the intersection labels}

\item{cexsn}{a character expansion for the set names}

\item{zcolor}{a vector of colors for the custom zones,
or predefined colors if "style"}

\item{...}{other parameters for the line plots.}
}
\examples{

}
