% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesExperiment-methods.R
\docType{methods}
\name{timepoints}
\alias{timepoints}
\alias{timepoints,TimeSeriesExperiment-method}
\alias{timepoints<-}
\alias{timepoints<-,TimeSeriesExperiment-method}
\title{Timepoint information}
\usage{
timepoints(object, ...)

\S4method{timepoints}{TimeSeriesExperiment}(object)

timepoints(object, ...) <- value

\S4method{timepoints}{TimeSeriesExperiment}(object) <- value
}
\arguments{
\item{object}{a \code{TimeSeriesExperiment} object.}

\item{...}{argiments to other functions.}

\item{value}{a numeric vector with new time information.}
}
\value{
a numeric vector
}
\description{
Timepoint information
}
\examples{
data("endoderm_small")
head(timepoints(endoderm_small))
timepoints(endoderm_small) <- sample(1:ncol(endoderm_small))
head(timepoints(endoderm_small))

}
