% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesExperiment-methods.R
\docType{methods}
\name{timeSeries}
\alias{timeSeries}
\alias{timeSeries,TimeSeriesExperiment-method}
\alias{timeSeries<-}
\alias{timeSeries<-,TimeSeriesExperiment-method}
\title{Time series formatted data.}
\usage{
timeSeries(object, ...)

\S4method{timeSeries}{TimeSeriesExperiment}(object, name = NULL)

timeSeries(object, ...) <- value

\S4method{timeSeries}{TimeSeriesExperiment}(object) <- value
}
\arguments{
\item{object}{a \code{TimeSeriesExperiment} object.}

\item{...}{argiments to other functions.}

\item{name}{a character string, one of 'ts', 'ts_with_lags', 'ts_collapsed'
and 'ts_collapsed_with_lags'. If NULL, all elements are returned.}

\item{value}{replacement list}
}
\value{
a \code{data.frame}
}
\description{
Getter and setter methods for \code{timeSeries} slot of
a \code{TimeSeriesExperiment} object.
}
\details{
\code{timeSeries} slot is a list with 'ts' and (optionally) 
'ts_collapsed' storing data formatted as time-series/time-courses.
}
\examples{
data("endoderm_small")
endoderm_small <- makeTimeSeries(endoderm_small)
head(timeSeries(endoderm_small))
head(timeSeries(endoderm_small, name = 'ts'))

}
