% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesExperiment-class.R
\docType{class}
\name{TimeSeriesExperiment-class}
\alias{TimeSeriesExperiment-class}
\alias{.TimeSeriesExperiment}
\alias{TimeSeriesExperiment-constructor}
\alias{TimeSeriesExperiment}
\title{Time Series Experiment Class}
\usage{
TimeSeriesExperiment(
  ...,
  timepoint = numeric(0),
  group = character(0),
  replicate = character(0)
)
}
\arguments{
\item{...}{For \link[SummarizedExperiment:RangedSummarizedExperiment-class]{
SummarizedExperiment:: SummarizedExperiment}, S4 methods \code{list} and 
\code{matrix}, arguments identical to those of the \code{SimpleList} method.}

\item{timepoint}{a vector indicating timepoint at which each sample 
was collected or a character string equal to one of the column names 
of a supplied colData.}

\item{group}{a vector indicating a group membership for each sample 
or a character string equal to one of the column names of a supplied 
colData. If not specified, the group is set to 'G1' for each sample.}

\item{replicate}{a vector indicating a replicate id of each sample 
or a character string equal to one of the column names of a supplied 
colData. If not specified, the replicate is set to 'R1' for each sample.}
}
\value{
Returns an initialized TimeSeriesExperiment object.
}
\description{
\code{TimeSeriesExperiment} is an extension of 
\code{SummarizedExperiment} class with the following new slots 
(in addition to SummarizedExperiment slots).

Constructor for 'TimeSeriesExperiment' object which stores
same data as \code{SummarizedExperiment} but also slots for time, group, 
replicate  and other time-series formated data useful for applyting 
data-analysis.
}
\details{
The \code{TimeSeriesExperiment} class is an object 
(the main data container) used for in the time series/course experiment 
analysis. It stores all relevant information associated with the 
dataset, including the raw data, group, replicate and time associated 
with each sample (column of the data). The object includes also slots 
for results from some class-specific methods.

\code{TimeSeriesExperiment} constructor initializes the 
\code{TimeSeriesExperiment} TimeSeriesExperiment object and populates
the time, replicate, and group slots.
}
\section{Slots}{

\describe{
\item{\code{timepoint}}{A vector indicating the time-point of each sample collection.}

\item{\code{group}}{A vector indicating the group membership of each sample.}

\item{\code{replicate}}{A vector indicating the replicate id of each sample.}

\item{\code{assayCollapsed}}{A matrix with assay data aggregated over replicates.}

\item{\code{colDataCollapsed}}{A \link{DataFrame} where rows correspond
to samples aggregated over replicates and columns indicate group 
membership and time-point.}

\item{\code{timeSeries}}{A list of time-course formatted data. Each element
of the list is a \link{DataFrame} with the first three columns 
indicating feature, group, replicate, and the remaining ones storing
the assay data at consecutive time points. \code{TimeSeriesExperiment} 
methods will typically generate elements of the list named: 'ts',
'ts_collapsed'.}

\item{\code{dimensionReduction}}{A list of results from applying 
dimmensionality reduction methods; elements named by technique used.}

\item{\code{clusterAssignment}}{A list of results from clustering of the time-series 
features (rows) containing elements: 'settings', 'hclust', 'clust_map'
and 'clust_centroids'.}

\item{\code{differentialExpression}}{A list of results from differential expression
analysis. Either from point-wise ('timepoint_de') or trajectory
('trajectory_de') differential expression analysis.}
}}

\examples{
raw <- matrix(runif(3000), ncol = 30)
timepoint <- rep(rep(1:5, each = 3), 2)
replicate <- rep(1:3, 10)
group <- rep(1:2, each = 15)
test_TimeSeriesExperiment <- TimeSeriesExperiment(
    assays = list(raw),
    timepoint = timepoint,
    replicate = replicate,
    group = group)
test_TimeSeriesExperiment

}
