% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timevis_init.R
\name{timevis_init}
\alias{timevis_init}
\title{Initialize and setup the timevis object}
\usage{
timevis_init(raw.data, project = "timevis time course project",
  sample.data = NULL, feature.data = NULL, time = NULL,
  time_column = NULL, replicate = NULL, replicate_column = FALSE,
  group = NULL, group_column = NULL)
}
\arguments{
\item{raw.data}{Raw input data with columns corresponding to samples
(observations) and rows to features}

\item{project}{Project name (string)}

\item{sample.data}{Optional. A \code{data.frame} object were rows are
samples (observations) and columns are sample attributes
(e.g. group/condition, replicate, time)}

\item{feature.data}{Optional. A \code{data.frame} object were rows are
features and columns are feature names under different conventions (e.g.
Enterez IDs,  gene symbols) or feature attributes (e.g. chromosome,
location in the genome, biotype, gc content etc.)}

\item{time}{Vector of length \code{ncol(raw.data)} indicating the time
at which the sample was collected. If not given, \code{time_column} MUST
be specified.}

\item{time_column}{A character string equal to one of the column names
of \code{sample.data}. This is an alternative, way of specifying the time
corresponding to samples, if \code{time} argument is not provided.}

\item{replicate}{Vector of length \code{ncol(raw.data)} indicating
the replicate from which the sample came from. If not given,
\code{replicate_column} will be used.}

\item{replicate_column}{A character string equal to one of the column names
of \code{sample.data}. This is an alternative, way of specifying the
replicate corresponding to samples, if \code{replicate} argument is not
provided. If both \code{replicate} and \code{replicate_column} are not set,
the function assumes all samples come from the same replicate, and assigns
a replicate name 'R1' to all samples.}

\item{group}{A character vector of length \code{ncol(raw.data)} indicating
the group membership of the sample. If not given, \code{group_column}
will be used.}

\item{group_column}{A character string equal to one of the column names
of \code{sample.data}. This is an alternative, way of specifying the
replicate corresponding to samples, if \code{replicate} argument is not
provided. If both \code{group} and \code{group_column} are not set, the
function assumes all samples come from the same group, and assigns a group
name 'G1' to all samples.}
}
\value{
Returns a timevis object with the raw data stored in object@raw.data.
object@group, object@replicate, and object@timealso initialized.
}
\description{
Initializes the timevis object and populates
the time, replicate, and group slots.
}
\examples{
endoderm_raw <- read.table(
  file = system.file('extdata', 'raw_endoderm_small.txt',
    package = 'timevis'),
  as.is = TRUE
)
time <- as.numeric(gsub("(.*)\\\\_D", "", colnames(endoderm_raw)))
replicate <- gsub("\\\\_(.*)", "", colnames(endoderm_raw))
group <- substring(replicate, 1, 1)
endoderm_small <- timevis(
  raw.data = endoderm_raw,
  time = time,
  replicate = replicate,
  group = group)
endoderm_small

}
