% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ts_adonis}
\alias{ts_adonis}
\title{Adonis for time series data.}
\usage{
ts_adonis(X, time, replicate = NULL, group = NULL, alpha_coeff = NULL,
  dist_method = "euclidean", p_adj_method = "BH", verbose = TRUE, ...)
}
\arguments{
\item{X}{a data matrix or data.frame where rows are time series.}

\item{time}{a vector of length equal to ncol(X) indicating the time
variable corresponding to the sample (data column).}

\item{replicate}{a vector of length equal to ncol(X) indicating
the sample replicate (e.g. individual).}

\item{group}{a vector of length equal to ncol(X) indicating
the group membership of the sample.}

\item{alpha_coeff}{weights for each lag difference, by default NULL.}

\item{dist_method}{the name of any method used in vegdist to calculate
pairwise distances, "euclidean" by defaults.}

\item{p_adj_method}{a correction method. Can be abbreviated.}

\item{...}{other options to adonis function.}
}
\value{
a data.frame with adonis results for all features.
}
\description{
A function computes distances between time series data
(with added lag differences) and uses vegan::adonis() function
to test for group differences.
}
