% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_class.R
\name{update_info}
\alias{update_info}
\title{check and update sample data}
\usage{
update_info(sample.data, time_column, replicate_column = NULL,
  group_column = NULL)
}
\arguments{
\item{sample.data}{A \code{data.frame} object were rows are samples
(observations) and columns are sample attributes 
(e.g. group/condition replicate, time)}

\item{time_column}{A character string equal to one of the column names
of \code{sample.data}. This is an alternative, way of specifying the time
corresponding to samples, if \code{time} argument is not provided.}

\item{replicate_column}{A character string equal to one of the column names
of \code{sample.data}. This is an alternative, way of specifying the
replicate corresponding to samples, if \code{replicate} argument is not
provided. If both \code{replicate} and \code{replicate_column} are not set,
the function assumes all samples come from the same replicate, and assigns
a replicate name 'R1' to all samples.}

\item{group_column}{A character string equal to one of the column names
of \code{sample.data}. This is an alternative, way of specifying the
replicate corresponding to samples, if \code{replicate} argument is not
provided. If both \code{group} and \code{group_column} are not set, the
function assumes all samples come from the same group, and assigns a group
name 'G1' to all samples.}
}
\value{
Returns a data frame.
}
\description{
Ensures validity of input sample data,
and updates the colums.
}
