% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{variance_stabilization}
\alias{variance_stabilization}
\title{Variance satbilization.}
\usage{
variance_stabilization(X, method = "log", pseudocount = 1,
  log.base = exp(1))
}
\arguments{
\item{X}{an assay data matrix or data.frame where columns correspond}

\item{method}{Method for variance stabilization
(VST). Currently, supports log plus one normalization, "log",
or DESeq2::varianceStabilizingTransformation, "deseq".
By default "lognorm" VST is performed.}

\item{pseudocount}{A numerical indicating the pseudocount in log
transformation}

\item{log.base}{A numerical indicating the base for log
transformation}
}
\value{
Returns a varianced stabilized data matrix.
}
\description{
This function performs variance stabilization
on assay data.
}
\examples{
X <- sapply(exp(rlnorm(10)), function(m) rnbinom(20, size = 1, mu = m))
head(X)
Y <- variance_stabilization(X, log.base = 2)
head(Y)

}
