% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{varianceStabilization}
\alias{varianceStabilization}
\title{Variance stabilization.}
\usage{
varianceStabilization(X, method = "asinh")
}
\arguments{
\item{X}{an assay data matrix or data.frame where columns correspond}

\item{method}{Method for variance stabilization (VST).
Currently, supports "none" (no VST), "log1p" (log plus one), "asinh"
(inverse hyperbolic sine) or "deseq"
(\code{\link[DESeq2]{varianceStabilizingTransformation}} function from
\code{DESeq2} package). Default is "log1p".}
}
\value{
Returns a varianced stabilized data matrix.
}
\description{
This function performs variance stabilization
on assay data.
}
\examples{
X <- sapply(exp(rlnorm(10)), function(m) rnbinom(20, size = 1, mu = m))
head(X)
Y <- varianceStabilization(X, method = "asinh")
head(Y)

}
