% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_class.R
\docType{class}
\name{vistimeseq}
\alias{vistimeseq}
\alias{vistimeseq-class}
\title{The Time Course Data Class}
\description{
The \code{vistimeseq} object is the main object in the time course
experiment analysis. It stores all relevant information associated with
the dataset, including the raw data, group, replicate and time associated
with each sample (column of the data). The object includes also slots for
results from some class-specific methods.

Each vistimeseq object has a number of key slots listed below.
}
\section{Slots}{

\describe{
\item{\code{project.name}}{A character string indicating the project name.}

\item{\code{raw.data}}{Raw data with columns corresponding to samples
(observations) and rows to features.}

\item{\code{norm.data}}{Normalized data.}

\item{\code{sample.data}}{A \code{data.frame} object were rows are samples
(observations) and columns are sample attributes (e.g. group/condition,
replicate, time)}

\item{\code{sample.data.collapsed}}{A \code{data.frame} with rows corresponding
to samples aggregated over replicates. The columns indicate group membership
and time.}

\item{\code{feature.data}}{A \code{data.frame} object were rows are features and
columns are feature names under different conventions (e.g. Enterez IDs,
gene symbols) or feature attributes (e.g. chromosome, location in the
genome, biotype, gc content etc.)}

\item{\code{group}}{A character vector of length \code{ncol(raw.data)} indicating
the group membership of the sample}

\item{\code{replicate}}{Vector of length \code{ncol(raw.data)} indicating
the replicate from which the sample came from}

\item{\code{time}}{Vector of length \code{ncol(raw.data)} indicating the time
at which the sample was collected}

\item{\code{timecourse.data}}{List of data in a time-course format. Each element
of the list is a \code{data.frame} where the data is organized in with
first three columns indicating feature, group, replicate, and the remaining
ones providing the data at each available time point.
\code{vistimeseq} methods will typically generate elements named:
'tc', 'tc_with_lags', 'tc_collapsed' and 'tc_collapsed_with_lags'.}

\item{\code{dim.red}}{List of stored dimmensional reductions; named by technique}

\item{\code{cluster.map}}{A \code{data.frame} storing results of gene clustering}

\item{\code{diff.expr}}{A \code{data.frame} storing results of differential
expression analysis}
}}

