% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_class.R
\name{vistimeseq_from_SummarizedExperiment}
\alias{vistimeseq_from_SummarizedExperiment}
\title{vistimeseq object and constructor from SummarizedExperiment}
\usage{
vistimeseq_from_SummarizedExperiment(sumexp, time_column,
  replicate_column = NULL, group_column = NULL,
  project = "'vistimeseq' time course project")
}
\arguments{
\item{sumexp}{SummarizedExperiment object}

\item{time_column}{A character string equal to one of the column names
of \code{sample.data}. This is an alternative, way of specifying the time
corresponding to samples, if \code{time} argument is not provided.}

\item{replicate_column}{A character string equal to one of the column names
of \code{sample.data}. This is an alternative, way of specifying the
replicate corresponding to samples, if \code{replicate} argument is not
provided. If both \code{replicate} and \code{replicate_column} are not set,
the function assumes all samples come from the same replicate, and assigns
a replicate name 'R1' to all samples.}

\item{group_column}{A character string equal to one of the column names
of \code{sample.data}. This is an alternative, way of specifying the
replicate corresponding to samples, if \code{replicate} argument is not
provided. If both \code{group} and \code{group_column} are not set, the
function assumes all samples come from the same group, and assigns a group
name 'G1' to all samples.}

\item{project}{Project name (string)}
}
\value{
Returns a vistimeseq object with the raw data stored in
object@raw.data, object@sample.data object@group, object@replicate,
and object@time are also initialized.
}
\description{
Initializes the vistimeseq object from SummarizedExperiment
and populates the time, replicate, and group slots.
}
\examples{
library(SummarizedExperiment)
raw <- matrix(runif(3000), ncol = 30)
pheno.data <- data.frame(
   time = rep(rep(1:5, each = 3), 2),
   replicate = rep(1:3, 10),
   group = rep(1:2, each = 15))
feature.data <- data.frame(
   feature = paste0("F", 1:100)
)
test_sumexp <- SummarizedExperiment(
   assays = list(counts = raw),
   rowData = feature.data, colData = pheno.data)
   
test_vistimeseq <- vistimeseq_from_SummarizedExperiment(
   test_sumexp, time_column = "time", group_column = "group",
   replicate_column = "replicate")
test_vistimeseq
}
