% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_lags}
\alias{add_lags}
\title{Add differences to the time-course data.}
\usage{
add_lags(object, lambda = c(0.5, 0.25))
}
\arguments{
\item{object}{A \code{vistimeseq} object}

\item{lambda}{Weights for each lag difference, for time-course data.
Length of \code{lambda} specifies number of lags to include. By default
lag of order one and two are included with coefficients 0.5 and 0.25
respectively.}
}
\value{
Returns \code{vistimeseq} object with lags added to elements
in \code{timecourse.data} slot.
}
\description{
This function concatenates lags to time-course
data stored in elements of the \code{timecourse.data} slot.
}
\examples{
data("endoderm_small")
endoderm_small <- normalize_data(endoderm_small)
endoderm_small <- convert_to_timecourse(endoderm_small)
endoderm_small <- add_lags(endoderm_small)
head(time_course(endoderm_small, collapsed = FALSE))
head(time_course(endoderm_small, collapsed = TRUE))

endoderm_small <- collapse_replicates(endoderm_small)
endoderm_small <- convert_to_timecourse(endoderm_small)
endoderm_small <- add_lags(endoderm_small)
head(time_course(endoderm_small, collapsed = TRUE))

}
