% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_lags}
\alias{add_lags}
\title{Add differences to the time-course data.}
\usage{
add_lags(object, lambda_coeff = c(0.5, 0.25))
}
\arguments{
\item{object}{A \code{timvis} object}

\item{lambda_coeff}{the weights for each lag difference,
the length of alphs_coeff specifies the degrees of lags
to include. By default lag one and lag two included
with coefficients equal to 0.5 and 0.25.}
}
\value{
Returns \code{timvis} object with
lags added to elements in \code{timecourse.data}
slot.
}
\description{
This function concatenates lags to time-course
data stored in elements of the \code{timecourse.data} slot.
}
\examples{
endoderm_small
endoderm_small <- normalize_data(endoderm_small)
endoderm_small <- collapse_replicates(endoderm_small)
endoderm_small <- convert_to_timecourse(endoderm_small)
endoderm_small <- add_lags(endoderm_small)
head(endoderm_small@timecourse.data$tc)

}
