% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_lags_to_tc}
\alias{add_lags_to_tc}
\title{Add differences to the time-course data.}
\usage{
add_lags_to_tc(timecourse, lambda_coeff)
}
\arguments{
\item{timecourse}{a data matrix or data.frame where each
column corresponds to consecutive time point.}

\item{lambda}{the weights for each lag difference,
the length of alphs_coeff specifies the degrees of lags
to include.}
}
\value{
a data matrix with added difference lags.
}
\description{
This function add lag difference features
to the time-course data.
}
\examples{
X <- matrix(rnorm(1000), ncol = 50)
group <- rep(c("A", "B"), each = 25)
replicate <- rep(paste0("rep", 1:5), each = 5)
time <- rep(1:5, 10)
tc <- data_to_tc(X, time, replicate, group)
tc_with_lags <- add_lags(tc[, 4:ncol(tc)], alpha_coeff = c(0.5, 0.25))
tc_with_lags <- cbind(tc[, 1:3], tc_with_lags)
head(tc_with_lags)

}
