% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_data.R
\name{add_sample_data}
\alias{add_sample_data}
\title{Add Featuredata}
\usage{
add_sample_data(object, sampledata, col.name = colnames(sampledata))
}
\arguments{
\item{object}{vistimeseq object}

\item{sampledata}{Data frame where the row names are sample names
and the columns are additional sample attributes}

\item{col.name}{Column names in sampledata to add to the object}
}
\value{
\code{vistimeseq} object where the additional sample data has been
added as columns in object@sample.data
}
\description{
Adds additional data for samples to \code{vistimeseq} object.
Can be any piece of information associated with a sample (e.g. subject
clinical or nonclinical data). The additional data can be used in
plotting functions.
}
\examples{
endoderm_small
idx <- sample(1:26, length(sample_names(endoderm_small)), replace = TRUE)
random_letters <- data.frame(
  letters = LETTERS[idx],
  row.names = sample_names(endoderm_small)
  )
endoderm_small <- add_sample_data(
  object = endoderm_small,
  sampledata = random_letters
  )
head(sample_data(endoderm_small))

}
