% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_data}
\alias{cluster_data}
\title{Cluster time series data.}
\usage{
cluster_data(X, dist = "euclidean", dynamic = FALSE,
  hclust_params = list(), static_cut_params = list(h = 0.5),
  dynamic_cut_params = list(max_height = 0.9))
}
\arguments{
\item{X}{a data matrix or data.frame where rows are time series.}

\item{dist}{the distance metric for the dissimilarity used for clustering.}

\item{dynamic}{whether dynamic branch cutting should be done for cluster
assignment.}

\item{hclust_params}{parameters for \code{\link[stats]{hclust}} function.}

\item{static_cut_params}{parameters for \code{\link{assign_cluster_static}}.}

\item{dynamic_cut_params}{parameters for \code{\link{assign_cluster_dynamic}}.}
}
\value{
a list with the \code{hclust} object, as well as \code{clust_map}
and \code{clust_centroids} data.frames.
}
\description{
Perform cluster assignment using hierarchical clustering
and static or dynamic branch cutting.
If 'subset' is specified the clsutering is performed on the subset
of the data, and the rest of the rows are assigned based on the distances
to the centroids of computed clusters.
}
\examples{
X <- get_data(endoderm_small)
clust_res <- cluster_data(X)
head(clust_res$clust_centroids)
head(clust_res$clust_map)

}
