% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_timecourse}
\alias{convert_to_timecourse}
\title{Convert data to time-course.}
\usage{
convert_to_timecourse(object, feature.trans.method = "var_stab",
  var.stabilize.method = "log1p")
}
\arguments{
\item{object}{A \code{vistimeseq} object}

\item{feature.trans.method}{Method for feature normalization. 
Default "none". Currently supports only "none" (no transformation),
"scale_feat_sum" (scaling by feature sum),
or "var_stab" (variance stabilization). Default is "var_stab".}

\item{var.stabilize.method}{Method for variance stabilization (VST).
Currently, supports "none" (no VST), "log1p" (log plus one), "asinh"
(inverse hyperbolic sine) or "deseq"
(\code{\link[DESeq2]{varianceStabilizingTransformation}} function from
\code{DESeq2} package). Default is "log1p".}
}
\value{
Returns \code{vistimeseq} object after conversion to
time-course format. Converted data is stored in
\code{timecourse.data} slot.
}
\description{
This function converts the wide data matrix
to time-course long \code{data.frame} format where each
row gives data values over time (at each time point) for each
feature, group, and replicate.
}
\examples{
data("endoderm_small")
endoderm_small <- normalize_data(endoderm_small)
endoderm_small <- convert_to_timecourse(endoderm_small)
head(time_course(endoderm_small))

}
