% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_methods.R
\docType{methods}
\name{get_dim_reduced}
\alias{get_dim_reduced}
\alias{get_dim_reduced,vistimeseq-method}
\title{Accessors for the 'dim.red' slot of a \code{vistimeseq} object.}
\usage{
get_dim_reduced(object, type = "all")

\S4method{get_dim_reduced}{vistimeseq}(object, type = "all")
}
\arguments{
\item{object}{a \code{vistimeseq} object.}

\item{type}{one of elements of 'dim.red' slot: 'pca_sample', 'pca_feature'
and 'pca_eigs' or 'all' for returning the entire list.}
}
\value{
a \code{data.frame} or a list of \code{data.frame}s
}
\description{
'dim.red' slots is a list with 'pca_sample', 'pca_feature' and
'pca_eigs' results from PCA projection.
}
\examples{
data("endoderm_small")
endoderm_small <- run_pca(endoderm_small)
head(get_dim_reduced(endoderm_small, "pca_sample")[, 1:3])

}
