% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{endoderm}
\alias{endoderm}
\alias{endoderm_small}
\title{Prices of 50,000 round cut diamonds.}
\format{A \code{vistimeseq} object}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE98411}
}
\usage{
data(endoderm)

endoderm_small
}
\description{
A `vistimeseq` object contating gene expression data, feature and sample
data from a study by Blake et al. "A Comparative Study Of Endoderm
Differentiation In Humans And Chimpanzees"
}
\references{
Blake et al. (2017)
(\href{https://www.biorxiv.org/content/early/2017/05/09/135442}{bioRxiv})
}
\keyword{datasets}
