% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{endoderm}
\alias{endoderm}
\alias{endoderm_small}
\title{Prices of 50,000 round cut diamonds.}
\format{A \code{vistimeseq} object}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE98411}
}
\usage{
data(endoderm)

endoderm_small
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds. The variables are as follows:
}
\details{
\itemize{
  \item price.
  \item carat.
}
}
\references{
Blake et al. (2017)
(\href{https://www.biorxiv.org/content/early/2017/05/09/135442}{bioRxiv})
}
\keyword{datasets}
