% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_methods.R
\docType{methods}
\name{feature_data}
\alias{feature_data}
\alias{feature_data,vistimeseq-method}
\alias{feature_data}
\alias{feature_data<-}
\alias{feature_data<-,vistimeseq-method}
\title{Accessors for the 'feature.data' slot of a \code{vistimeseq} object.}
\usage{
feature_data(object)

\S4method{feature_data}{vistimeseq}(object)

feature_data(object) <- value

\S4method{feature_data}{vistimeseq}(object) <- value
}
\arguments{
\item{object}{a \code{vistimeseq} object.}

\item{value}{a \code{data.frame} with new feature data}
}
\value{
a \code{data.frame}
}
\description{
'feature.data' slots holds information on individual features including
corresponding feature name in column 'feature'.
}
\examples{
endoderm_small
head(feature_data(endoderm_small))

}
