% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Plot selected time series.}
\usage{
plot_time_series(object, features = feature_names(object), smooth = TRUE,
  ncol = 5)
}
\arguments{
\item{object}{A \code{vistimeseq} object.}

\item{features}{A vector of names of selected features to plot.}

\item{smooth}{If TRUE a smoothed line is plotted for each gene
and each group, else a piecewise linear average (over replicates) curve
is plotted.}

\item{ncol}{An integer indicating the number of columns for facetting.
Default is 5.}
}
\value{
list of ggplot objects
}
\description{
Plotting expression over time for selected genes curve
and colors correspond to distinct groups.
}
