% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_ts_clusters}
\alias{plot_ts_clusters}
\title{Plot (time) series over clusters.}
\usage{
plot_ts_clusters(object, features = NULL, transparency = 0.5, ncol = 4)
}
\arguments{
\item{object}{A \code{vistimeseq} object.}

\item{features}{A vector of names of selected features to plot.}

\item{transparency}{transparency of trajectory lines.}

\item{ncol}{number of columns in the facte plot.}
}
\value{
ggplot object
}
\description{
Plots timecourse (aggregated over replicates) feature data
faceted by computed clusters and experimental group.
}
\examples{
endoderm_small <- cluster_timecourse_features(endoderm_small)
plot_ts_clusters(endoderm_small)

}
