% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_methods.R
\docType{methods}
\name{project_name}
\alias{project_name}
\alias{project_name,vistimeseq-method}
\alias{project_name}
\alias{project_name<-}
\alias{project_name<-,vistimeseq-method}
\title{Project name}
\usage{
project_name(object)

\S4method{project_name}{vistimeseq}(object)

project_name(object) <- value

\S4method{project_name}{vistimeseq}(object) <- value
}
\arguments{
\item{object}{a \code{vistimeseq} object.}

\item{value}{a character string}
}
\value{
a character string
}
\description{
Returns project name.
}
\examples{
endoderm_small
project_name(endoderm_small)

}
