% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_methods.R
\docType{methods}
\name{get_replicate}
\alias{get_replicate}
\alias{set_replicate<-}
\alias{get_replicate}
\alias{set_replicate}
\title{Accessors for the 'replicate' slot of a \code{vistimeseq} object.}
\usage{
get_replicate(object)

set_replicate(object) <- value

\S4method{get_replicate}{vistimeseq}(object)

\S4method{set_replicate}{vistimeseq}(object) <- value
}
\arguments{
\item{object}{a \code{vistimeseq} object.}
}
\description{
'grop' slots holds feature names
}
\examples{

endoderm_small
head(get_replicate(endoderm_small))

}
