% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{run_pca}
\alias{run_pca}
\title{Run PCA scores.}
\usage{
run_pca(object, collapse.replicates = FALSE, group.selected = NULL,
  var.stabilize.method = "log1p")
}
\arguments{
\item{object}{A \code{vistimeseq} object}

\item{collapse.replicates}{Whether PCA should be computed on the data
with replicates aggregated.}

\item{group.selected}{An optional character string indicating a particular
group of samples PCA should be applied to. By default all groups are
included.}

\item{var.stabilize.method}{Method for variance stabilization (VST).
Currently, supports "none" (no VST), "log1p" (log plus one), "asinh"
(inverse hyperbolic sine) or "deseq"
(\code{\link[DESeq2]{varianceStabilizingTransformation}} function from
\code{DESeq2} package). Default is "log1p".}
}
\value{
Returns \code{vistimeseq} object with PCA results in \code{dim.red}
slot, a lists containing matrices of coordinates 'pca_sample', and
'pca_features', as well as a vector 'pca_eigs'.
}
\description{
Compute the projection of either samples or time series
features onto a PCA space. In case, of PCA for features, the PCA can be
computed for individual sample group as indicated or for samples from all
groups. In either, case the data is first collapsed over replicates, so
that each gene is represented as a vector of a single time course.
}
\examples{
endoderm_small
endoderm_small <- run_pca(endoderm_small)
head(endoderm_small@dim.red$pca_sample[, 1:5])
head(endoderm_small@dim.red$pca_eigs)

}
