% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_methods.R
\docType{methods}
\name{sample_data}
\alias{sample_data}
\alias{sample_data,vistimeseq-method}
\alias{sample_data}
\alias{sample_data<-}
\alias{sample_data<-,vistimeseq-method}
\title{Accessors for the 'sample.data' slot of a \code{vistimeseq} object.}
\usage{
sample_data(object)

\S4method{sample_data}{vistimeseq}(object)

sample_data(object) <- value

\S4method{sample_data}{vistimeseq}(object) <- value
}
\arguments{
\item{object}{a \code{vistimeseq} object}

\item{value}{a \code{data.frame} with new sample data}
}
\description{
'sample.data' slots holds information on individual samples including
corresponding sample name in column 'sample' as well as time, group
and replicate.
}
\examples{

endoderm_small
head(sample_data(endoderm_small))

}
