% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesExperiment-methods.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,TimeSeriesExperiment-method}
\title{show method for \code{TimeSeriesExperiment}}
\usage{
\S4method{show}{TimeSeriesExperiment}(object)
}
\arguments{
\item{object}{A TimeSeriesExperiment object}
}
\value{
nothing, just prints to console
}
\description{
show method for \code{TimeSeriesExperiment}
}
