% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_class.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,vistimeseq-method}
\title{show method for vistimeseq}
\usage{
\S4method{show}{vistimeseq}(object)
}
\arguments{
\item{object}{A vistimeseq object}
}
\description{
show method for vistimeseq
}
