% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timevis_class.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,timevis-method}
\title{show method for timevis}
\usage{
\S4method{show}{timevis}(object)
}
\arguments{
\item{object}{A timevis object}
}
\description{
show method for timevis
}
