% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_methods.R
\docType{methods}
\name{get_time}
\alias{get_time}
\alias{get_time,vistimeseq-method}
\alias{set_time<-}
\alias{set_time<-,vistimeseq-method}
\title{Accessors for the 'time' slot of a \code{vistimeseq} object.}
\usage{
get_time(object)

\S4method{get_time}{vistimeseq}(object)

set_time(object) <- value

\S4method{set_time}{vistimeseq}(object) <- value
}
\arguments{
\item{object}{a \code{vistimeseq} object.}

\item{value}{a numeric vector with new time information.}
}
\value{
a numeric vector
}
\description{
'grop' slots holds feature names
}
\examples{
endoderm_small
head(get_time(endoderm_small))

}
