% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_class.R
\name{vistimeseq}
\alias{vistimeseq}
\alias{vistimeseq-class}
\alias{vistimeseq}
\title{The Time Course Data Class}
\usage{
vistimeseq(raw.data, project = "'vistimeseq' time course project",
  sample.data = NULL, feature.data = NULL, time = NULL,
  time_column = NULL, replicate = NULL, replicate_column = NULL,
  group = NULL, group_column = NULL)

vistimeseq(raw.data, project = "'vistimeseq' time course project",
  sample.data = NULL, feature.data = NULL, time = NULL,
  time_column = NULL, replicate = NULL, replicate_column = NULL,
  group = NULL, group_column = NULL)
}
\arguments{
\item{raw.data}{Raw input data with columns corresponding to samples
(observations) and rows to features}

\item{project}{Project name (string)}

\item{sample.data}{Optional. A \code{data.frame} object were rows are
samples (observations) and columns are sample attributes
(e.g. group/condition, replicate, time)}

\item{feature.data}{Optional. A \code{data.frame} object were rows are
features and columns are feature names under different conventions (e.g.
Enterez IDs,  gene symbols) or feature attributes (e.g. chromosome,
location in the genome, biotype, gc content etc.)}

\item{time}{Vector of length \code{ncol(raw.data)} indicating the time
at which the sample was collected. If not given, \code{time_column} MUST
be specified.}

\item{time_column}{A character string equal to one of the column names
of \code{sample.data}. This is an alternative, way of specifying the time
corresponding to samples, if \code{time} argument is not provided.}

\item{replicate}{Vector of length \code{ncol(raw.data)} indicating
the replicate from which the sample came from. If not given,
\code{replicate_column} will be used.}

\item{replicate_column}{A character string equal to one of the column names
of \code{sample.data}. This is an alternative, way of specifying the
replicate corresponding to samples, if \code{replicate} argument is not
provided. If both \code{replicate} and \code{replicate_column} are not set,
the function assumes all samples come from the same replicate, and assigns
a replicate name 'R1' to all samples.}

\item{group}{A character vector of length \code{ncol(raw.data)} indicating
the group membership of the sample. If not given, \code{group_column}
will be used.}

\item{group_column}{A character string equal to one of the column names
of \code{sample.data}. This is an alternative, way of specifying the
replicate corresponding to samples, if \code{replicate} argument is not
provided. If both \code{group} and \code{group_column} are not set, the
function assumes all samples come from the same group, and assigns a group
name 'G1' to all samples.}
}
\value{
Returns a vistimeseq object with the raw data stored in
object@raw.data, object@sample.data object@group, object@replicate,
and object@time are also initialized.
}
\description{
The \code{vistimeseq} object is the main object in the time course
experiment analysis. It stores all relevant information associated with
the dataset, including the raw data, group, replicate and time associated
with each sample (column of the data). The object includes also slots for
results from some class-specific methods.

Each vistimeseq object has a number of key slots listed below.

Initializes the vistimeseq object and populates
the time, replicate, and group slots.
}
\section{Slots}{

\describe{
\item{\code{project.name}}{A character string indicating the project name.}

\item{\code{raw.data}}{Raw data with columns corresponding to samples
(observations) and rows to features.}

\item{\code{norm.data}}{Normalized data.}

\item{\code{sample.data}}{A \code{data.frame} object were rows are samples
(observations) and columns are sample attributes (e.g. group/condition,
replicate, time)}

\item{\code{sample.data.collapsed}}{A \code{data.frame} with rows corresponding
to samples aggregated over replicates. The columns indicate group membership
and time.}

\item{\code{feature.data}}{A \code{data.frame} object were rows are features and
columns are feature names under different conventions (e.g. Enterez IDs,
gene symbols) or feature attributes (e.g. chromosome, location in the
genome, biotype, gc content etc.)}

\item{\code{group}}{A character vector of length \code{ncol(raw.data)} indicating
the group membership of the sample}

\item{\code{replicate}}{Vector of length \code{ncol(raw.data)} indicating
the replicate from which the sample came from}

\item{\code{time}}{Vector of length \code{ncol(raw.data)} indicating the time
at which the sample was collected}

\item{\code{timecourse.data}}{List of data in a time-course format. Each element
of the list is a \code{data.frame} where the data is organized in with
first three columns indicating feature, group, replicate, and the remaining
ones providing the data at each available time point.
\code{vistimeseq} methods will typically generate elements named:
'tc', 'tc_with_lags', 'tc_collapsed' and 'tc_collapsed_with_lags'.}

\item{\code{dim.red}}{List of stored dimmensional reductions; named by technique}

\item{\code{cluster.features}}{A list storing results from timecourse (feature)
clustering.}

\item{\code{diff.expr}}{A list storing results of differential expression analysis.}
}}

\examples{
raw <- matrix(runif(3000), ncol = 30)
time <- rep(rep(1:5, each = 3), 2)
replicate <- rep(1:3, 10)
group <- rep(1:2, each = 15)
test_vistimeseq <- vistimeseq(
raw.data = raw,
time = time,
replicate = replicate,
group = group)
test_vistimeseq

}
