% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistimeseq_methods.R
\docType{methods}
\name{get_data}
\alias{get_data}
\alias{get_data,vistimeseq-method}
\alias{get_data}
\alias{set_raw_data<-}
\alias{set_raw_data<-,vistimeseq-method}
\alias{set_data<-}
\alias{set_data<-,vistimeseq-method}
\title{Accessors for the 'raw.data' and 'data' slots of a \code{vistimeseq}
object.}
\usage{
get_data(object, raw = FALSE)

\S4method{get_data}{vistimeseq}(object, raw = FALSE)

set_raw_data(object) <- value

\S4method{set_raw_data}{vistimeseq}(object) <- value

set_data(object) <- value

\S4method{set_data}{vistimeseq}(object) <- value
}
\arguments{
\item{object}{a \code{vistimeseq} object.}

\item{raw}{whether raw data should be returned}

\item{value}{a data matrix or \code{data.frame}}
}
\description{
The 'raw.data' of 'data' slots hold data where columns are
samples taken at specific timepoints and rows are features. If normalization
was performed 'data' holds normalized data
}
\examples{
endoderm_small
head(get_data(endoderm_small))

}
